will only call setSize and setPreferredSize if number of columns did not change; solves bug in #29 (solves #29)

+++ src/main/java/de/uib/configed/gui/hwinfopage/PanelHWInfo.java
@@ -287,8 +287,8 @@ public class PanelHWInfo extends JPanel implements TreeSelectionListener {
 			mult = 1024;
 		}
 		// TODO: nano, micro
-		if (v.compareTo(BigInteger.valueOf(mult * mult * mult)) >= 0) {
-			return ((float) Math.round(v.floatValue() * 1000 / (mult * mult * mult)) / 1000) + " G" + unit;
+		if (v.compareTo(BigInteger.valueOf((long) mult * mult * mult)) >= 0) {
+			return ((float) Math.round(v.floatValue() * 1000 / ((long) mult * mult * mult)) / 1000) + " G" + unit;
 		} else if (v.compareTo(BigInteger.valueOf(mult * mult)) >= 0) {
 			return ((float) Math.round(v.floatValue() * 1000 / (mult * mult)) / 1000) + " M" + unit;
 		} else if (v.compareTo(BigInteger.valueOf(mult)) >= 0) {
